+++ app/Exceptions/Service/Database/NoSuitableDatabaseHostException.php
@@ -1,0 +1,16 @@
+<?php
+
+namespace Pterodactyl\Exceptions\Service\Database;
+
+use Pterodactyl\Exceptions\DisplayException;
+
+class NoSuitableDatabaseHostException extends DisplayException
+{
+    /**
+     * NoSuitableDatabaseHostException constructor.
+     */
+    public function __construct()
+    {
+        parent::__construct('No database host was found that meets the requirements for this server.');
+    }
+}