+++ CHANGELOG.md
@@ -18,6 +18,7 @@ This project follows [Semantic Versioning](http://semver.org) guidelines.
 * Adds back client API for sending commands or power toggles to a server though the Panel API: `/api/client/servers/<identifier>`
 * Added proper transformer for Packs and re-enabled missing includes on server.
 * Added support for using Filesystem as a caching driver, although not recommended.
+* Added support for user management of server databases.
 
 ## v0.7.3 (Derelict Dermodactylus)
 ### Fixed